requirements.txt
Flask
SpeechRecognition
mysql-connector-python
pydub
torchaudio
torch
speechbrain
soundfile

